+++ tests/unit/keyless.test.ts
@@ -2,18 +2,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { keylessTestConfig, keylessTestObject } from "./helper";
-import { Deserializer, Hex, KeylessAccount, KeylessPublicKey, KeylessSignature } from "../../src";
+import {
+  Deserializer,
+  Hex,
+  KeylessAccount,
+  KeylessPublicKey,
+  KeylessSignature,
+  verifyKeylessSignatureWithJwkAndConfig,
+} from "../../src";
 
 describe("Keyless", () => {
   describe("keylessPublicKey", () => {
-    it("should create the instance correctly without error", () => {
-      // Create from inputs
+    it("should create the instance correctly without error", async () => {
       const publicKey = new KeylessPublicKey(keylessTestObject.iss, keylessTestObject.idCommitment);
       expect(publicKey).toBeInstanceOf(KeylessPublicKey);
       expect(publicKey.toString()).toEqual(keylessTestObject.publicKey);
 
-      // Create from JWT and pepper
-      const publicKey2 = KeylessPublicKey.fromJwtAndPepper({
+      const publicKey2 = await KeylessPublicKey.fromJwtAndPepper({
         jwt: keylessTestObject.JWT,
         pepper: keylessTestObject.pepper,
       });
@@ -21,29 +26,28 @@ describe("Keyless", () => {
       expect(publicKey2.toString()).toEqual(keylessTestObject.publicKey);
     });
 
-    it("should verify the signature correctly", () => {
+    it("should verify the signature correctly", async () => {
       const publicKey = new KeylessPublicKey(keylessTestObject.iss, keylessTestObject.idCommitment);
       const signature = KeylessSignature.deserialize(
         new Deserializer(Hex.hexInputToUint8Array(keylessTestObject.signatureHex)),
       );
 
-      // Convert message to hex
       const hexMsg = Hex.fromHexString(keylessTestObject.messageEncoded);
 
-      // Verify with correct signed message
-      expect(
-        publicKey.verifySignature({
+      await expect(
+        verifyKeylessSignatureWithJwkAndConfig({
+          publicKey,
           message: hexMsg.toUint8Array(),
           signature,
           jwk: keylessTestObject.jwk,
           keylessConfig: keylessTestConfig,
         }),
-      ).toBe(true);
+      ).resolves.not.toThrow();
     });
   });
 
   describe("keylessAccount", () => {
-    it("should sign and verify a message correctly", () => {
+    it("should sign and verify a message correctly", async () => {
       const account = KeylessAccount.create({
         jwt: keylessTestObject.JWT,
         pepper: keylessTestObject.pepper,
@@ -53,22 +57,16 @@ describe("Keyless", () => {
       const message = "hello";
       const signature = account.sign(message);
       expect(signature).toBeInstanceOf(KeylessSignature);
-      expect(
-        account.publicKey.verifySignature({
+
+      await expect(
+        verifyKeylessSignatureWithJwkAndConfig({
+          publicKey: account.publicKey,
           message,
           signature,
           jwk: keylessTestObject.jwk,
           keylessConfig: keylessTestConfig,
         }),
-      ).toBe(true);
-      expect(
-        account.verifySignature({
-          message,
-          signature,
-          jwk: keylessTestObject.jwk,
-          keylessConfig: keylessTestConfig,
-        }),
-      ).toBe(true);
+      ).resolves.not.toThrow();
     });
   });
 });